* fixed resume bug from previous commit * corrected usages of portable types and format strings

+++ disk_file_posix.cpp
@@ -74,7 +74,7 @@ public:
 		return n;
 	}
 
-	int size()
+	uint64_t size()
 	{
 		// TODO
 		#error "Not implemented"
